% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTLD}
\alias{getTLD}
\title{Retrieve a current list of the top level domains}
\usage{
getTLD(url = "https://publicsuffix.org/list/effective_tld_names.dat")
}
\arguments{
\item{url}{URL of the location for the tld name authority}
}
\description{
This function will reach out to \url{https://publicsuffix.org/list/effective_tld_names.dat},
retrieve the contents and create a simple vector of all the top level domains (removing
comments and blanks lines from the file).
}
\details{
If there is no network connectivity, a cached version of the data ("tldnames") is included with
this package and can be loaded with \code{data("tldnames")} after loading this package.
}
\examples{
\dontrun{
tldnames <- getTLD()
}
}

