% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tldextract}
\alias{tldextract}
\title{Extract the top level domain, domain and subdomain from a host name}
\usage{
tldextract(host, tlds = NULL)
}
\arguments{
\item{host}{vector of one or more host names}

\item{tlds}{vector of TLD names (see \code{\link{getTLD}})}
}
\description{
Given one or more host names, this will return a data frame with four column:
\itemize{
\item "host": the original host name
\item "tld": the top level domain extracted
\item "domain" the domain extracted
\item "subdomain" one or more subdomains prepending the domain
}
}
\details{
If a hostname is not understandable (no top level domain is matched), it will
return NA for the three components.
}
\examples{
\dontrun{
hosts <- tldextract(c("www.google.com", "testing.co.uk"), tlds=getTLD())
}
}
\seealso{
getTLD
}

