% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{tldnames}
\alias{tldnames}
\title{List of Top-Level Domains Names}
\format{A vector containing the top level domains}
\description{
A dataset containing a single vector of the top-level domain names as retrevied from
\url{https://publicsuffix.org/list/effective_tld_names.dat} on August 2, 2014.
}
\details{
This can manually refreshed upon library load with the \code{\link{getTLD}} function.

Note, the non-ASCII characters may create a problem, and needs more testing.
}
\keyword{datasets}

