% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tldextract.R
\name{getTLD}
\alias{getTLD}
\title{Retrieve a current list of the top level domains}
\usage{
getTLD(url = "https://publicsuffix.org/list/effective_tld_names.dat")
}
\arguments{
\item{url}{URL of the location for the tld name authority. Set to point to publicsuffix.org
by default; while you can change this, much of \code{getTLD}'s internal parsing may be
specific to publicsuffix.org's format, and so other URLs may actively break or produce
oddly formated results}
}
\description{
This function will reach out to \url{https://publicsuffix.org/list/effective_tld_names.dat},
retrieve the contents and create a simple vector of all the top level domains (removing
comments and blanks lines from the file).

If there is no network connectivity, a cached version of the data ("tldnames") is included with
this package and can be loaded with \code{data("tldnames")} after loading this package.
}
\examples{
\dontrun{
tldnames <- getTLD()
}
}

